Support highlighting misfolding regions in puzzlemaker
